Extract traffic lights into component
